+++ src/main/java/com/example/taximotoapp_backend/User/service/UserService.java
@@ -77,12 +77,12 @@ public class UserService implements UserDetailsService {
         );
     }
 
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Utilisateur introuvable"));
     }
 
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
